+++ lib/flserial.dart
@@ -35,12 +35,14 @@ class SerialConfig {
   int dataBits; // 5, 6, 7, 8
   int stopBits; // 1, 2
   int parity; // 0: none, 1: odd, 2: even
+  int flowControl; // 0: none, 1: RTS/CTS, 2: XON/XOFF
 
   SerialConfig({
     this.baudRate = 115200,
     this.dataBits = 8,
     this.stopBits = 1,
     this.parity = 0,
+    this.flowControl = 0,
   });
 }
 
@@ -113,6 +115,7 @@ class FlSerial {
         config.dataBits,
         config.stopBits,
         config.parity,
+        config.flowControl,
       );
 
       if (!success) {